feat: Gaming: organize code #12
