+++ src/Objects/Common/UpdateScoreLabel.js
