+++ src/Scenes/GamePlay/SnakeAll.js
@@ -3,11 +3,11 @@ import { range, randomInt } from "201flaviosilva-utils";
 import GlobalConfigs from "../../Configs";
 import DebugFPS from "../../Components/DebugFPS";
 import Button from "../../Components/Button";
-import EventSystem, { EVENTS_NAMES } from "../../Objects/Snake/EventSystem";
+import EventSystem, { EVENTS_NAMES } from "../../Objects/Common/EventSystem";
 
 import { TextStyle } from "../../Theme";
 
-import UpdateScoreLabel from "../../Objects/Snake/UpdateScoreLabel";
+import UpdateScoreLabel from "../../Objects/Common/UpdateScoreLabel";
 import Player from "../../Objects/Snake/Player";
 import Apple from "../../Objects/Snake/Apple";
